move files and types around to prepare for Routing
