+++ src/pages/app/events.ts
@@ -1,4 +1,11 @@
-import { AppEvent } from "~app/types"
+export enum AppEvent {
+  MOUNTED = "e-mounted",
+  SELECTION_CHANGED = "e-selection-changed",
+  FOLDER_CHANGED = "e-folder-changed",
+  RENAME_NODE = "e-rename-node",
+  REMOVE_NODES = "e-remove-nodes",
+  CREATE_NODE = "e-create-node",
+}
 
 export const dispatch = (
   element: HTMLElement | EventTarget,
