+++ src/pages/app/explorer/node/node-icon.ts
@@ -1,11 +1,10 @@
-import { NodeType } from "~app/types"
 import fileImage from "~images/file.png"
 import folderImage from "~images/folder.png"
 import fullFolderImage from "~images/full_folder.png"
 import textFileImage from "~images/text_file.png"
 import audioFileImage from "~images/audio_file.png"
 import unknownImage from "~images/unknown.png"
-import { Node, state } from "~app/state"
+import { Node, NodeType, state } from "~pages/app/state"
 
 export function nodeIcon(node: Node): string {
   switch (node.type) {
